+++ sentry-android-integration-tests/sentry-uitest-android/src/androidTest/java/io/sentry/uitest/android/EnvelopeTests.kt
@@ -13,6 +13,7 @@ import io.sentry.SentryEvent
 import io.sentry.SentryOptions
 import io.sentry.protocol.SentryTransaction
 import org.junit.runner.RunWith
+import java.io.File
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertTrue
@@ -59,6 +60,32 @@ class EnvelopeTests : BaseUiTest() {
                 assertTrue(profilingTraceData.transactionName == "e2etests")
                 assertTrue(profilingTraceData.environment.isNotEmpty())
                 assertTrue(profilingTraceData.cpuArchitecture.isNotEmpty())
+            }
+            assertNoOtherEnvelopes()
+            assertNoOtherRequests()
+        }
+    }
+
+    @Test
+    fun checkProfileNotSentIfEmpty() {
+
+        initSentry(true) { options: SentryOptions ->
+            options.tracesSampleRate = 1.0
+            options.profilesSampleRate = 1.0
+        }
+        relayIdlingResource.increment()
+        val transaction = Sentry.startTransaction("e2etests", "test empty")
+        transaction.finish()
+        // Let's modify the trace file to be empty, so that the profile will actually be empty.
+        val profilesDirPath = Sentry.getCurrentHub().options.profilingTracesDirPath
+        val origProfileFile = File(profilesDirPath!!).listFiles()?.maxByOrNull { f -> f.lastModified() }
+        origProfileFile?.writeBytes(ByteArray(0))
+
+        relay.assert {
+            assertEnvelope {
+                it.assertItem<SentryTransaction>()
+                // Since the profile is empty, it is discarded and not sent to Sentry
+                it.assertNoOtherItems()
             }
             assertNoOtherEnvelopes()
             assertNoOtherRequests()